saved color in user defaults

+++ 10-EmojiArt/EmojiArt/ColorChooser.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct ColorChooser: View {
     @ObservedObject var document: EmojiArtDocument
     @Binding var chosenColor: Color
-    @Binding var chosenAlpha: Float
+    @Binding var chosenAlpha: Double
     @State private var isColorEditorPresented = false
     var body: some View {
         Image(systemName: "circles.hexagongrid").imageScale(.large)
